pyproject.toml
[project]
name = "pyznuny"
version = "0.0.1"
description = "A Python client for interacting with the Znuny ticketing system API."
authors = [
    { name = "Junior Rosa", email = "jr.dasrosas@gmail.com" },
    { name = "Pablo Gascon", email = "pablogasconiel445@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
]

[project.urls]
Homepage = "https://github.com/Junior-Rosa/py-znuny"
Repository = "https://github.com/Junior-Rosa/py-znuny"
Issues = "https://github.com/Junior-Rosa/py-znuny/issues"

[dependency-groups]
dev = [
    "linkify-it-py>=2.0.3",
    "myst-parser>=5.0.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.13",
    "sphinx>=9.1.0",
    "sphinx-autodoc-typehints>=3.6.2",
    "sphinx-rtd-theme>=3.1.0",
]

[tool.setuptools.packages.find]
where = ["src"]


[tool.ruff]
line-length = 88
target-version = "py314"
select = ["E", "F", "I", "B"]

[tool.ruff.format]
quote-style = "double"
